// frontend/src/api/stub/models/BodyCreateOrderOrderPost.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OrderCreate } from './OrderCreate';
import {
    OrderCreateFromJSON,
    OrderCreateFromJSONTyped,
    OrderCreateToJSON,
    OrderCreateToJSONTyped,
} from './OrderCreate';
import type { OrderLineItemBase } from './OrderLineItemBase';
import {
    OrderLineItemBaseFromJSON,
    OrderLineItemBaseFromJSONTyped,
    OrderLineItemBaseToJSON,
    OrderLineItemBaseToJSONTyped,
} from './OrderLineItemBase';

/**
 * 
 * @export
 * @interface BodyCreateOrderOrderPost
 */
export interface BodyCreateOrderOrderPost {
    /**
     * 
     * @type {OrderCreate}
     * @memberof BodyCreateOrderOrderPost
     */
    order: OrderCreate;
    /**
     * 
     * @type {Array<OrderLineItemBase>}
     * @memberof BodyCreateOrderOrderPost
     */
    orderLineItems: Array<OrderLineItemBase>;
}

/**
 * Check if a given object implements the BodyCreateOrderOrderPost interface.
 */
export function instanceOfBodyCreateOrderOrderPost(value: object): value is BodyCreateOrderOrderPost {
    if (!('order' in value) || value['order'] === undefined) return false;
    if (!('orderLineItems' in value) || value['orderLineItems'] === undefined) return false;
    return true;
}

export function BodyCreateOrderOrderPostFromJSON(json: any): BodyCreateOrderOrderPost {
    return BodyCreateOrderOrderPostFromJSONTyped(json, false);
}

export function BodyCreateOrderOrderPostFromJSONTyped(json: any, ignoreDiscriminator: boolean): BodyCreateOrderOrderPost {
    if (json == null) {
        return json;
    }
    return {
        
        'order': OrderCreateFromJSON(json['order']),
        'orderLineItems': ((json['order_line_items'] as Array<any>).map(OrderLineItemBaseFromJSON)),
    };
}

export function BodyCreateOrderOrderPostToJSON(json: any): BodyCreateOrderOrderPost {
    return BodyCreateOrderOrderPostToJSONTyped(json, false);
}

export function BodyCreateOrderOrderPostToJSONTyped(value?: BodyCreateOrderOrderPost | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'order': OrderCreateToJSON(value['order']),
        'order_line_items': ((value['orderLineItems'] as Array<any>).map(OrderLineItemBaseToJSON)),
    };
}