// frontend/src/api/stub/models/OrderLineItemBase.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Schema for creating a new line item
 * @export
 * @interface OrderLineItemBase
 */
export interface OrderLineItemBase {
    /**
     * 
     * @type {number}
     * @memberof OrderLineItemBase
     */
    menuItemId: number;
    /**
     * 
     * @type {string}
     * @memberof OrderLineItemBase
     */
    itemName: string;
    /**
     * 
     * @type {number}
     * @memberof OrderLineItemBase
     */
    unitPrice: number;
    /**
     * 
     * @type {number}
     * @memberof OrderLineItemBase
     */
    quantity?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof OrderLineItemBase
     */
    modifierIds?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof OrderLineItemBase
     */
    notes?: string | null;
}

/**
 * Check if a given object implements the OrderLineItemBase interface.
 */
export function instanceOfOrderLineItemBase(value: object): value is OrderLineItemBase {
    if (!('menuItemId' in value) || value['menuItemId'] === undefined) return false;
    if (!('itemName' in value) || value['itemName'] === undefined) return false;
    if (!('unitPrice' in value) || value['unitPrice'] === undefined) return false;
    return true;
}

export function OrderLineItemBaseFromJSON(json: any): OrderLineItemBase {
    return OrderLineItemBaseFromJSONTyped(json, false);
}

export function OrderLineItemBaseFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderLineItemBase {
    if (json == null) {
        return json;
    }
    return {
        
        'menuItemId': json['menu_item_id'],
        'itemName': json['item_name'],
        'unitPrice': json['unit_price'],
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
        'modifierIds': json['modifier_ids'] == null ? undefined : json['modifier_ids'],
        'notes': json['notes'] == null ? undefined : json['notes'],
    };
}

export function OrderLineItemBaseToJSON(json: any): OrderLineItemBase {
    return OrderLineItemBaseToJSONTyped(json, false);
}

export function OrderLineItemBaseToJSONTyped(value?: OrderLineItemBase | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'menu_item_id': value['menuItemId'],
        'item_name': value['itemName'],
        'unit_price': value['unitPrice'],
        'quantity': value['quantity'],
        'modifier_ids': value['modifierIds'],
        'notes': value['notes'],
    };
}